Added the pointer "ssl" to association with the SSL session for a connection to the "inOSSL_data_s" struct. Added a pointer to a "hostent" structure to this struct as well.
Function prototypes/signatures in the header file where changed because
the names of some functions have changed. Added the new function to perform
a "secure" connection to a server via SSL (this is a function that both
connects to a server and performs SSL handshake/negotiation).

Several header files were included in the "inSSL.h" header file.

A changed was made on how a certificate is loaded into memory (I do not
remember exactly what the change is, but it appears to be a certificate
type simplification).

A cleanup was added to the creation of a server from files (I do not recall
what this is).

The function to create a server by using certificate/key from data inmemory
was added. (This is similar to creating one with files, but the data is in
memory. This is such that we can potentially embed the data without writing
a file, perhaps with some live operation of binary embedding.)

A variable (instead of hardwired) socket "backlog" was added to the server
listening socket creation.

Some more verbose diagnostics on querying a server's certificate (when
acting as a client) were added.

Non-blocking socket capability was added to the client connection to a
server (this allows for the client to be more flexible and dynamic). A
blocking socket is also supported, and the behaviour is dictated at the
function invocation via a flag.

The function "inOSSL_ConnectToServerSSL()" was added to allow for connecting
to a server via SSL with either blocking or non-blocking sockets. (This is
a big deal, because it is hard to understand by reading the half-assed manual
pages of OpenSSL how to do this. It also enables for the client to be very
dynamic.)

The driver code was altered to accommodate the changes made everywhere here.
IN 2021/12/20